fix: add ability to disable escape and force ${var} usage. (#69)
* fix: add ability to disable escape and force ${var} usage.

In order to support expands vars in a windows environment, where paths
may exist, as:

\foo\bar\

The \ should not be the escape char.

Related, add a strictness for the var symbol, requiring ${}. This is
optional, but useful to further qualify when a var should be
interpreted.

* fix: api updates from review

MR review requested api change.